# Conflicts:
#	.gitignore
#	public/admin_panel/application/controllers/Edit_c.php
#	public/admin_panel/application/views/edit/trainer.php
#	templates/base.html.twig

+++ src/Controller/UserController.php
@@ -3,7 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Teachers;
+use App\Entity\TrainerVideo;
 use App\Entity\User;
+use App\Service\TrainerVideoService;
+use App\Traits\EntityManager;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -14,12 +17,28 @@ use Symfony\Component\Security\Core\User\PasswordUpgraderInterface;
 
 class UserController extends AbstractController
 {
+    use EntityManager;
 
+    private const ACTION_ADD = 'add';
+
+    private const ACTION_DELETE = 'delete';
+
+    private const ACTION_ALL = 'all';
+
+    /**
+     * @var UserPasswordEncoderInterface
+     */
     private $passwordEncoder;
+
+    /**
+     * @var TrainerVideoService
+     */
+    private $trainerVideoService;
 
     public function __construct(UserPasswordEncoderInterface $passwordEncoder)
     {
         $this->passwordEncoder = $passwordEncoder;
+        $this->trainerVideoService = new TrainerVideoService($this->getEntityManager());
     }
 
     /**
@@ -27,19 +46,19 @@ class UserController extends AbstractController
      */
     public function index()
     {
-      if ($this->getUser()->getistrainer()) {
-        return $this->redirectToRoute('main');
-      } else {
-        return $this->render('templates/cabinet/user/timelist.html.twig',
-            [
-                'router' => 'cabinet',
-                'styles' => [
-                    'cabinet/cabinet.css',
-                    'cabinet/user/timelist.css'
+        if ($this->getUser()->getistrainer()) {
+            return $this->redirectToRoute('main');
+        } else {
+            return $this->render('templates/cabinet/user/timelist.html.twig',
+                [
+                    'router' => 'cabinet',
+                    'styles' => [
+                        'cabinet/cabinet.css',
+                        'cabinet/user/timelist.css'
+                    ]
                 ]
-            ]
-        );
-      }
+            );
+        }
     }
 
     /**
@@ -47,19 +66,19 @@ class UserController extends AbstractController
      */
     public function specifications()
     {
-      if ($this->getUser()->getistrainer()) {
-        return $this->redirectToRoute('trainer_index');
-      } else {
-        return $this->render('templates/cabinet/user/specifications.html.twig',
-            [
-                'router' => 'cabinet',
-                'styles' => [
-                    'cabinet/cabinet.css',
-                    'cabinet/user/specifications.css'
+        if ($this->getUser()->getistrainer()) {
+            return $this->redirectToRoute('trainer_index');
+        } else {
+            return $this->render('templates/cabinet/user/specifications.html.twig',
+                [
+                    'router' => 'cabinet',
+                    'styles' => [
+                        'cabinet/cabinet.css',
+                        'cabinet/user/specifications.css'
+                    ]
                 ]
-            ]
-        );
-      }
+            );
+        }
     }
 
     /**
@@ -67,44 +86,44 @@ class UserController extends AbstractController
      */
     public function history()
     {
-      if ($this->getUser()->getistrainer()) {
-        return $this->redirectToRoute('main');
-      } else {
-        $data = [
-            [
-                'date' => 'Сегодня, 25 июня',
-                'items' => [
-                    [
-                        'time' => '13:00-14:00',
-                        'image' => '/images/temp/Rectangle506.png',
-                        'name' => 'Rafaeeela',
-                        'rating' => 3,
-                        'price' => 3000,
-                        'closed' => true,
-                    ],
-                    [
-                        'time' => '13:00-14:00',
-                        'image' => '/images/temp/Rectangle506.png',
-                        'name' => 'Rafaeeela',
-                        'rating' => 3,
-                        'price' => 3000,
-                        'closed' => true,
+        if ($this->getUser()->getistrainer()) {
+            return $this->redirectToRoute('main');
+        } else {
+            $data = [
+                [
+                    'date' => 'Сегодня, 25 июня',
+                    'items' => [
+                        [
+                            'time' => '13:00-14:00',
+                            'image' => '/images/temp/Rectangle506.png',
+                            'name' => 'Rafaeeela',
+                            'rating' => 3,
+                            'price' => 3000,
+                            'closed' => true,
+                        ],
+                        [
+                            'time' => '13:00-14:00',
+                            'image' => '/images/temp/Rectangle506.png',
+                            'name' => 'Rafaeeela',
+                            'rating' => 3,
+                            'price' => 3000,
+                            'closed' => true,
+                        ]
                     ]
                 ]
-            ]
-        ];
+            ];
 
-        return $this->render('templates/cabinet/user/history.html.twig',
-            [
-                'router' => 'cabinet',
-                'styles' => [
-                    'cabinet/cabinet.css',
-                    'cabinet/user/history.css'
-                ],
-                'data' => $data,
-            ]
-        );
-      }
+            return $this->render('templates/cabinet/user/history.html.twig',
+                [
+                    'router' => 'cabinet',
+                    'styles' => [
+                        'cabinet/cabinet.css',
+                        'cabinet/user/history.css'
+                    ],
+                    'data' => $data,
+                ]
+            );
+        }
     }
 
     /**
@@ -113,17 +132,17 @@ class UserController extends AbstractController
     public function settings()
     {
         if ($this->getUser()->getistrainer()) {
-          return $this->redirectToRoute('main');
+            return $this->redirectToRoute('main');
         } else {
-          return $this->render('templates/cabinet/user/settings.html.twig',
-              [
-                  'router' => 'cabinet',
-                  'styles' => [
-                      'cabinet/cabinet.css',
-                      'cabinet/user/settings.css'
-                  ]
-              ]
-          );
+            return $this->render('templates/cabinet/user/settings.html.twig',
+                [
+                    'router' => 'cabinet',
+                    'styles' => [
+                        'cabinet/cabinet.css',
+                        'cabinet/user/settings.css'
+                    ]
+                ]
+            );
         }
     }
 
@@ -143,7 +162,7 @@ class UserController extends AbstractController
 
         if (!$user) {
             throw $this->createNotFoundException(
-                'No user found for id '.$id
+                'No user found for id ' . $id
             );
         }
 
@@ -159,16 +178,15 @@ class UserController extends AbstractController
             'family' => $user->getFamily(),
             'discord' => $user->getDiscord(),
             'purse' => $user->getPurse(),
+            'vk' => $user->getVk(),
         ];
 
-        if($user->getIsTrainer())
-        {
+        if ($user->getIsTrainer()) {
             $trainer = $this->getDoctrine()
                 ->getRepository(Teachers::class)
                 ->findByUserId(intval($id));
 
-            if(!$trainer)
-            {
+            if (!$trainer) {
                 $entityManager = $this->getDoctrine()->getManager();
                 /** @var Teachers $trainer */
                 $trainer = new Teachers();
@@ -185,21 +203,11 @@ class UserController extends AbstractController
                 $entityManager->persist($trainer);
                 $entityManager->flush();
             }
-
 //            $trainer->setUser(null);
             $user->setTrainer($trainer);
 
             $userTrainer = $user->getTrainer();
             $result['trainer'] = $trainer;
-//                [
-//                'videolink' => $user->getVideoLink(),
-//                'cost' => $trainer->getCost(),
-//                'about' => $trainer->getAbout(),
-//                'shorttitle' => $trainer->getShorttitle(),
-//                'method' => $trainer->getMethod(),
-//                'streamtype' => $trainer->getStreamType(),
-//                'twitch' => $trainer->getTwitch(),
-//            ];
         }
 
         return $this->json($result);
@@ -210,16 +218,6 @@ class UserController extends AbstractController
      */
     public function setUserInfo(Request $request)
     {
-//        /** @var User $client */
-//        $client = $this->get('security.token_storage')->getToken()->getUser();
-//
-//
-//        if($client->getId() != $userId)
-//        {
-//            $this->json("U have not access to update this user");
-//        }
-////        $newInfo = json_decode($form);
-
         // retrieves POST variables respectively\
         $userId = $request->request->get('id');
         $entityManager = $this->getDoctrine()->getManager();
@@ -231,7 +229,7 @@ class UserController extends AbstractController
                 'status' => 404
             ]);
         }
-
+        $user->setVk($request->request->get('vk'));
         $user->setName($request->request->get('name'));
         $user->setEmail($request->request->get('email'));
         $user->setNickname($request->request->get('nickname'));
@@ -241,46 +239,15 @@ class UserController extends AbstractController
         $user->setRank($request->request->get('rank'));
         $user->setDiscord($request->request->get('discord'));
 //
-        if($request->request->get('photo') && ($request->request->get('photo') != ''))
-        {
-            if ($user->getPhoto() !== null)
-            {
-                if(($user->getPhoto() == $user->getId() . "-" . $user->getName() . $user->getNickname() . ".jpg" )||
-                ($user->getPhoto() == $user->getId() . "-" . $user->getName() . $user->getNickname() . ".png" )||
-                ($user->getPhoto() == $user->getId() . "-" . $user->getName() . $user->getNickname() . ".jpeg" ))
-                {
+        if ($request->request->get('photo') && ($request->request->get('photo') != '')) {
+            if ($user->getPhoto() !== null) {
+                if (($user->getPhoto() == $user->getId() . "-" . $user->getName() . $user->getNickname() . ".jpg") ||
+                    ($user->getPhoto() == $user->getId() . "-" . $user->getName() . $user->getNickname() . ".png") ||
+                    ($user->getPhoto() == $user->getId() . "-" . $user->getName() . $user->getNickname() . ".jpeg")) {
                     unlink("images/temp/matches/" . $user->getPhoto());
                 }
 
             }
-            // split the string on commas
-            // $data[ 0 ] == "data:image/png;base64"
-            // $data[ 1 ] == <actual base64 string>
-//            $data = explode( ',', $request->request->get('photo') );
-//            switch ($data[0])
-//            {
-//                case "data:image/png;base64":
-//                    $phototype = "png";
-//                    break;
-//                case "data:image/jpeg;base64":
-//                    $phototype = "jpeg";
-//                    break;
-//                case "data:image/jpg;base64":
-//                    $phototype = "jpg";
-//                    break;
-//                default:
-//                    $phototype = "jpeg";
-//                    break;
-//            }
-////            $phototype = $request->request->get('phototype');
-//            $filename = $user->getId() . "-" . $user->getName() . $user->getNickname() . "." . $phototype;
-//            $output = $data[1];
-//            $output_file_name = "\images\\temp\matches\\" . $filename;
-//            $ifp = fopen( $output_file_name, 'wb');// stream_get_contents(fopen( $output_file_name, 'wb'));//open( $output_file, 'wb' );
-//
-//            $image = base64_decode($output);
-//            fwrite( $ifp, $image );
-//            fclose( $ifp );
 
             $data = $request->request->get('photo');
 
@@ -288,7 +255,7 @@ class UserController extends AbstractController
                 $data = substr($data, strpos($data, ',') + 1);
                 $type = strtolower($type[1]); // jpg, png, gif
 
-                if (!in_array($type, [ 'jpg', 'jpeg', 'png' ])) {
+                if (!in_array($type, ['jpg', 'jpeg', 'png'])) {
                     throw new \Exception('invalid image type');
                 }
 
@@ -304,39 +271,37 @@ class UserController extends AbstractController
             $filename = $user->getId() . "-" . $user->getName() . $user->getNickname() . "." . $type;
             file_put_contents("images/temp/matches/{$filename}", $data);
 
-            $user->setPhoto( $filename);
+            $user->setPhoto($filename);
         }
 
 
-        if($user->getIsTrainer())
-        {
+        if ($user->getIsTrainer()) {
             $trainer = $entityManager->getRepository(Teachers::class)
                 ->findOneBy([
                     'userid' => $userId
-                    ]);
-            if(!$trainer)
-            {
+                ]);
+            if (!$trainer) {
                 $trainer = new Teachers();
                 $trainer->setUser($user->getId());
             }
 
+            $trainer->setTimeZone($request->request->get('timezone', null));
             $trainer->setVideoLink($request->request->get('videolink'));
             $trainer->setCost($request->request->get('cost'));
             $trainer->setAbout($request->request->get('about'));
             $trainer->setShorttitle($request->request->get('shorttitle'));
             $trainer->setMethod($request->request->get('method'));
-            $trainer->setStreamType($request->request->get('streamtype'));
+            $trainer->setStreamType($request->request->get('streamtype', null));
             $trainer->setTwitch($request->request->get('channellink'));
+            $trainer->setIsLessonCost($request->request->get('isLessonCost', false) === "false" ? false: true);
 
             $entityManager->persist($trainer);
             $entityManager->flush();
         }
         $password = $request->request->get('password');
 
-        if ($password == $request->request->get('password_repeat'))
-        {
-            if($password != '')
-            {
+        if ($password == $request->request->get('password_repeat')) {
+            if ($password != '') {
                 $user->setPassword($this->passwordEncoder->encodePassword($user, $password));
             }
         }
@@ -348,8 +313,6 @@ class UserController extends AbstractController
             'status' => 200,
             'photo' => $user->getPhoto()
         ]);
-
-//        return $this->json($user);
     }
 
     /**
@@ -374,4 +337,36 @@ class UserController extends AbstractController
         return $this->json($user);
     }
 
+    /**
+     * @Route("/ru/{action}/user/video/{id}", methods={"POST"}, name="add_trainer_video")
+     */
+    public function addVideo(Request $request, $action, $id)
+    {
+        $request = json_decode($request->getContent());
+        $trainer = $this->getDoctrine()->getRepository(User::class)->find($id);
+        $videos = $this->trainerVideoService->getByTrainer($trainer);
+
+        if ($action === self::ACTION_ADD and isset($trainer) and count($videos) < TrainerVideoService::MAX_COUNT) {
+            $isYouTube = $this->trainerVideoService->isYouTubeVideo($request->video);
+
+            if ($isYouTube) {
+                $this->trainerVideoService->create($trainer, $request->video);
+            } else {
+                return $this->json([
+                    'video' => 'Ссылка на видео должны быть с Youtube! В формате https://www.youtube.com/qwda2fg'
+                ], 422);
+            }
+        }
+        if ($action === self::ACTION_DELETE and isset($trainer)) {
+            $this->trainerVideoService->deleteVideo($trainer, $request->video);
+        }
+        $trainerVideos = $this->trainerVideoService->getByTrainer($trainer);
+
+        $videos = [];
+        /** @var TrainerVideo $trainerVideo */
+        foreach ($trainerVideos as $trainerVideo) {
+            $videos[] = $trainerVideo->getVideoUrl();
+        }
+        return $this->json($videos);
+    }
 }
